+++ Schema/generateSchema.ps1
@@ -1,9 +1,14 @@
 $generatorPath = "C:\src\mseng\AppInsights-Common"
 $schemasPath = "C:\src\mseng\DataCollectionSchemas"
 $publicSchemaLocation = "https://raw.githubusercontent.com/Microsoft/ApplicationInsights-Home/sergkanz/schemas/EndpointSpecs/Schemas/Bond"
+$localPublicSchema = $true
 
 
 $currentDir = $scriptPath = split-path -parent $MyInvocation.MyCommand.Definition
+#fix path
+$generatorPath = "$generatorPath\..\bin\Debug\BondSchemaGenerator\BondSchemaGenerator"
+$schemasPath = "$schemasPath\v2\Bond\"
+
 
 
 function RegExReplace([string]$fileName, [string]$regex, [string]$replacement="")
@@ -14,45 +19,58 @@ function RegExReplace([string]$fileName, [string]$regex, [string]$replacement=""
     del $tempFileName
 }
 
-#fix path
-$generatorPath = "$generatorPath\..\bin\Debug\BondSchemaGenerator\BondSchemaGenerator"
-$schemasPath = "$schemasPath\v2\Bond\"
 
-& "$generatorPath\BondSchemaGenerator.exe" -v -i "$schemasPath\AppInsightsTypes.bond" -i "$schemasPath\PerformanceCounterData.bond" -i "$schemasPath\SessionStateData.bond" -i "$schemasPath\ContextTagKeys.bond" -o "$currentDir\PublicSchema\" -e BondLanguage -t BondLayout -n test --flatten false
+#####################################################################
+## PUBLIC SCHEMA
+#####################################################################
 
-#mkdir -Force $currentDir\PublicSchema
+mkdir -Force $currentDir\PublicSchema
 
-@(
-"Base.bond",
-"ContextTagKeys.bond",
-"Data.bond", 
-"DataPoint.bond", 
-"DataPointType.bond", 
-"Domain.bond", 
-"Envelope.bond", 
-"EventData.bond", 
-"ExceptionData.bond", 
-"ExceptionDetails.bond", 
-"MessageData.bond", 
-"MetricData.bond", 
-"PageViewData.bond", 
-"PageViewPerfData.bond", 
-"RemoteDependencyData.bond", 
-"RequestData.bond", 
-"SeverityLevel.bond", 
-"StackFrame.bond"
-)  | ForEach-Object { 
+del "$currentDir\PublicSchema\*.bond"
 
-    $fileName = $_
-    #& Invoke-WebRequest -o "$currentDir\PublicSchema\$fileName" "$publicSchemaLocation/$fileName"
+if ($localPublicSchema) {
+    # Generate public schema using bond generator
+    & "$generatorPath\BondSchemaGenerator.exe" -v -i "$schemasPath\AppInsightsTypes.bond" -i "$schemasPath\PerformanceCounterData.bond" -i "$schemasPath\SessionStateData.bond" -i "$schemasPath\ContextTagKeys.bond" -o "$currentDir\PublicSchema\" -e BondLanguage -t BondLayout -n test --flatten false
+} else {
+    # Download public schema from the github
+    @(
+    "Base.bond",
+    "ContextTagKeys.bond",
+    "Data.bond", 
+    "DataPoint.bond", 
+    "DataPointType.bond", 
+    "Domain.bond", 
+    "Envelope.bond", 
+    "EventData.bond", 
+    "ExceptionData.bond", 
+    "ExceptionDetails.bond", 
+    "MessageData.bond", 
+    "MetricData.bond", 
+    "PageViewData.bond", 
+    "PageViewPerfData.bond", 
+    "RemoteDependencyData.bond", 
+    "RequestData.bond", 
+    "SeverityLevel.bond", 
+    "StackFrame.bond"
+    )  | ForEach-Object { 
+        $fileName = $_
+        & Invoke-WebRequest -o "$currentDir\PublicSchema\$fileName" "$publicSchemaLocation/$fileName"
+    }
 }
 
+
+#####################################################################
+## BOND-GENERATED CODE
+#####################################################################
+
 mkdir -Force $currentDir\obj
+
 Invoke-WebRequest -o "$currentDir\obj\nuget.exe" https://api.nuget.org/downloads/nuget.exe
 
 
-& "$currentDir\obj\nuget" install Bond.CSharp -Version 4.2.1 -OutputDirectory "$currentDir\obj\packages"
+del $currentDir\obj\gbc\*
 
+& "$currentDir\obj\nuget" install Bond.CSharp -Version 4.2.1 -OutputDirectory "$currentDir\obj\packages"
 
 dir "$currentDir\PublicSchema" | ForEach-Object { 
     & "$currentDir\obj\packages\Bond.CSharp.4.2.1\tools\gbc.exe" c# --collection-interfaces --using="DateTimeOffset=System.DateTimeOffset" --using="TimeSpan=System.TimeSpan" --using="Guid=System.Guid" -o "$currentDir\obj\gbc" $_.FullName
@@ -62,11 +80,20 @@ del "$currentDir\obj\gbc\*_interfaces.cs"
 del "$currentDir\obj\gbc\*_services.cs"
 del "$currentDir\obj\gbc\*_proxies.cs"
 
+
+
+#####################################################################
+## CLEAR BOND-GENERATED CODE OUT OF BOND REFERENCES
+#####################################################################
+
+
 dir "$currentDir\obj\gbc" | ForEach-Object { 
     # Rename namespace from AI to Microsoft.ApplicationInsights.Extensibility.Implementation.External
     RegExReplace $_.FullName "(namespace AI)" "namespace Microsoft.ApplicationInsights.Extensibility.Implementation.External"
+    RegExReplace $_.FullName "new Dictionary" "new ConcurrentDictionary"
     # Remove "using Bond" statements
     RegExReplace $_.FullName "using Bond.*"
+    RegExReplace $_.FullName "using System.Collections.Generic;" "using System.Collections.Concurrent;`r`n    using System.Collections.Generic;"
     # Remove all Bond attributes
     RegExReplace $_.FullName "\[global::Bond\..*\]"
     # Remove derivations from Microsoft.Telemetry.Domain
@@ -85,32 +112,14 @@ dir "$currentDir\obj\gbc" | ForEach-Object {
     RegExReplace $_.FullName "= nothing;" "= null;"
 }
 
-@(
-"AvailabilityData_types",
-"Base_types",
-"ContextTagKeys_types",
-"DataPointType_types",
-"DataPoint_types",
-"Data_types",
-"Domain_types",
-"DependencyKind_types",
-"DependencySourceType_types",
-"Envelope_types",
-"EventData_types",
-"ExceptionData_types",
-"ExceptionDetails_types",
-"MessageData_types",
-"MetricData_types",
-"PageViewData_types",
-"PerformanceCounterData_types",
-"RemoteDependencyData_types",
-"RequestData_types",
-"SeverityLevel_types",
-"SessionState_types"
-"SessionStateData_types",
-"StackFrame_types",
-"TestResult_types"
-) | ForEach-Object { 
+
+#####################################################################
+## COPY GENERATED FILES TO THE REPOSITORY
+#####################################################################
+
+del "$currentDir\..\src\Core\Managed\Shared\Extensibility\Implementation\External\*_types.cs"
+
+dir "$currentDir\obj\gbc\*_types.cs" | ForEach-Object { 
     $fileName = $_
-    copy "$currentDir\obj\gbc\$fileName.cs" "$currentDir\..\src\Core\Managed\Shared\Extensibility\Implementation\External\"
+    copy $fileName "$currentDir\..\src\Core\Managed\Shared\Extensibility\Implementation\External\"
 }