+++ Test/CoreSDK.Test/Shared/TelemetryClientTest.cs
@@ -346,7 +346,7 @@
             var dependency = (DependencyTelemetry)sentTelemetry.Single();
 
             Assert.Equal("name", dependency.Name);
-            Assert.Equal("command name", dependency.CommandName);
+            Assert.Equal("command name", dependency.Data);
             Assert.Equal(timestamp, dependency.Timestamp);
             Assert.Equal(TimeSpan.FromSeconds(42), dependency.Duration);
             Assert.Equal(false, dependency.Success);
@@ -611,7 +611,9 @@
             var configuration = new TelemetryConfiguration { TelemetryChannel = channel, InstrumentationKey = "Test Key" };
             var client = new TelemetryClient(configuration);
 
+#pragma warning disable 618
             Assert.DoesNotThrow(() => client.Track(new SessionStateTelemetry()));
+#pragma warning disable 618
         }
 
         [TestMethod]
@@ -838,9 +840,13 @@
             ExceptionTelemetry telemetry4 = new ExceptionTelemetry();
             MetricTelemetry telemetry5 = new MetricTelemetry();
             PageViewTelemetry telemetry6 = new PageViewTelemetry();
+#pragma warning disable 618
             PerformanceCounterTelemetry telemetry7 = new PerformanceCounterTelemetry();
+#pragma warning restore 618
             RequestTelemetry telemetry8 = new RequestTelemetry();
+#pragma warning disable 618
             SessionStateTelemetry telemetry9 = new SessionStateTelemetry();
+#pragma warning restore 618
             TraceTelemetry telemetry10 = new TraceTelemetry();
             AvailabilityTelemetry telemetry11 = new AvailabilityTelemetry();
 
@@ -860,7 +866,7 @@
             };
 
             // ChuckNorrisTeamUnitTests resource in Prototypes5
-            var config = new TelemetryConfiguration { InstrumentationKey = "fafa4b10-03d3-4bb0-98f4-364f0bdf5df8" };
+            var config = new TelemetryConfiguration { InstrumentationKey = "cdec18df-21e6-42a0-b7d2-3e84d5793d97" };
             var telemetryClient = new TelemetryClient(config);
 
             telemetryClient.Initialize(telemetry1);
@@ -879,7 +885,7 @@
 
             HttpClient client = new HttpClient();
             var result = client.PostAsync(
-                "https://dc.services.visualstudio.com/v2/validate",
+                "https://eus-breeze-in.cloudapp.net/v2/validate",
                 new ByteArrayContent(Encoding.UTF8.GetBytes(json))).GetAwaiter().GetResult();
 
             if (result.StatusCode != HttpStatusCode.OK)