+++ src/Core/Managed/Net46/Extensibility/Implementation/RichPayloadEventSource.cs
@@ -126,21 +126,23 @@
                     telemetryItem.Data,
                     Keywords.Exceptions);
             }
+#pragma warning disable 618
             else if (item is PerformanceCounterTelemetry)
             {
-                if (!this.EventSourceInternal.IsEnabled(EventLevel.Verbose, Keywords.PerformanceCounters))
+                if (!this.EventSourceInternal.IsEnabled(EventLevel.Verbose, Keywords.Metrics))
                 {
                     return;
                 }
 
-                var telemetryItem = item as PerformanceCounterTelemetry;
+                var telemetryItem = (item as PerformanceCounterTelemetry).Data;
                 this.WriteEvent(
-                    PerformanceCounterTelemetry.TelemetryName,
+                    MetricTelemetry.TelemetryName,
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.Tags,
                     telemetryItem.Data,
-                    Keywords.PerformanceCounters);
+                    Keywords.Metrics);
             }
+#pragma warning restore 618
             else if (item is PageViewTelemetry)
             {
                 if (!this.EventSourceInternal.IsEnabled(EventLevel.Verbose, Keywords.PageViews))
@@ -156,20 +158,21 @@
                     telemetryItem.Data,
                     Keywords.PageViews);
             }
+#pragma warning disable 618
             else if (item is SessionStateTelemetry)
             {
-                if (!this.EventSourceInternal.IsEnabled(EventLevel.Verbose, Keywords.SessionState))
+                if (!this.EventSourceInternal.IsEnabled(EventLevel.Verbose, Keywords.Events))
                 {
                     return;
                 }
 
-                var telemetryItem = item as SessionStateTelemetry;
+                var telemetryItem = (item as SessionStateTelemetry).Data;
                 this.WriteEvent(
-                    SessionStateTelemetry.TelemetryName,
+                    EventTelemetry.TelemetryName,
                     telemetryItem.Context.InstrumentationKey,
                     telemetryItem.Context.Tags,
                     telemetryItem.Data,
-                    Keywords.SessionState);
+                    Keywords.Events);
             }
             else if (item is AvailabilityTelemetry)
             {
@@ -264,16 +267,6 @@
             /// Keyword for page views.
             /// </summary>
             public const EventKeywords PageViews = (EventKeywords)0x40;
-
-            /// <summary>
-            /// Keyword for performance counters.
-            /// </summary>
-            public const EventKeywords PerformanceCounters = (EventKeywords)0x80;
-
-            /// <summary>
-            /// Keyword for session state.
-            /// </summary>
-            public const EventKeywords SessionState = (EventKeywords)0x100;
 
             /// <summary>
             /// Keyword for availability.